Started Encode Implementation
Encode implemented, uses functions calculateValue, calculateParity, and
fillSymbol
calculateValue implemented
calculateParity needs to be implemented. Should look at last (X) bytes,
based on bit depth, and calculate what the value for the parity symbol
should be.
fillSymbol needs to be implemented. This should draw a rectangle in the
right place on the bitmap and fill it with the passed in value.

May need to blank the bitmap to a predefined value after it's created.